`ConcurrentIter::has_more` is implemented
Not all iterators have this information exactly due to multiple reasons.
*  `ExactSizeConcurrentIter`s know exactly how many elements are there to yield. Therefore, they can give an exact answer.
* `ConcurrentIter`s built out of non-exact size `Iterator`s, such as iterators with a `filter`, does not exactly know how many more elements to be iterated over.
  * If the underlying iterator has not yet ended, it might or might not yield any more elements.
  * However, when the underlying iterator terminated, it can give the exact "no" answer.

In either case, the concurrent iterator is able to provide an exact "no" answer. This answer is already very useful for parallel processing, which could prevent spawning unnecessary threads.

An idea could be to return the following enum:

```rust
enum HasMore {
    Yes(usize),
    Maybe,
    No,
}
```

Closes #32