vec into con iter safety
Memory leak issue of concurrent iterators made from a vec when the iterator is not fully consumed is fixed. As the core method of batch iterators, `take_slice` method now returns `NoLeakIter` which iterates over all remaining elements on drop, guaranteeing that all items taken out of the original vector will be properly dropped.

Further, the underlying data structure of `ConIterOfVec` is updated. Now the vec is immediately destroyed on transformation and pointer is kept instead. `drop` and consuming `into_seq_iter` methods are revised accordingly.

Tests are extended so that it covers all possible ways that the data can be dropped:
* as an unused concurrent iter
* as a partially progressed concurrent iter
* as a concurrent iter that is immediately transformed into a sequential iter
* as a partially progressed concurrent iter transformed into a sequential iter which is partially progressed or completed