ConIterOfVec early exit safety
Memory leak issue of early exits on consuming iterators on vectors is fixed. Now the counter atomically maxes the count while returning the number of taken elements. If a thread observes that there are items to yield, that thread is responsible for dropping the elements.

Drop tests are extended to include early exits.