ConIterOfIter is simplified
In the prior version, the concurrent state of the concurrent iterator of an arbitrary Iterator was maintained by three atomic fields: reserved_counter, yielded_counter and state. In this PR, state is removed and the concurrent state is handled by the two counters. This does not have a significant impact on performance; however, makes the struct smaller and simplifies the implementation.

In addition a benchmark for testing the performance of ConIterOfIter is added. Results are promising. We especially get significant performance improvements when using a buffered iterator and iterating in chunks. Note that currently this is not possible with rayon as `par_bridge` does not seem to have the `chunks` method. Whenever/if it becomes available, this rayon variant must be added to the benchmark as well.